+++ flutter/test/screenshot/recorder_test.dart
@@ -9,7 +9,6 @@ import 'dart:ui';
 import 'package:flutter/widgets.dart' as widgets;
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
-import 'package:sentry_flutter/src/replay/replay_recorder.dart';
 import 'package:sentry_flutter/src/screenshot/recorder.dart';
 import 'package:sentry_flutter/src/screenshot/recorder_config.dart';
 import 'package:sentry_flutter/src/screenshot/screenshot.dart';
@@ -112,31 +111,6 @@ void main() async {
       });
 
     expect(await fixture.capture(), isNull);
-  });
-
-  // TODO: remove in the next major release, see _SentryFlutterExperimentalOptions.
-  group('Widget filter is used based on config or application', () {
-    test('Uses widget filter by default for Replay', () {
-      final sut = ReplayScreenshotRecorder(
-        ScreenshotRecorderConfig(),
-        defaultTestOptions(),
-      );
-      expect(sut.hasWidgetFilter, isTrue);
-    });
-
-    test('Does not use widget filter by default for Screenshots', () {
-      final sut =
-          ScreenshotRecorder(ScreenshotRecorderConfig(), defaultTestOptions());
-      expect(sut.hasWidgetFilter, isFalse);
-    });
-
-    test(
-        'Uses widget filter for Screenshots when privacy configured explicitly',
-        () {
-      final sut = ScreenshotRecorder(ScreenshotRecorderConfig(),
-          defaultTestOptions()..experimental.privacy.maskAllText = false);
-      expect(sut.hasWidgetFilter, isTrue);
-    });
   });
 
   group('$Screenshot', () {